Factor event dispatch out into separate file
